Initial Commit
This is a really sloppy commit. I didn't know what to omit so I'm
dumping everything in here, save for obvious exclusions like cache and
solution files.

Some of these things might be machine-specific, so I apologize for
committing them. Unfortunately I don't know what's required due to the
fact that I have a build data included.

If/When you merge Warren's code, make sure to only grab his scripts and
prefabs as well as their metadata.

We might want to look into a .gitignore that's useful for Unity
projects.